remove duplicated context
