+++ packages/nextjs/context/nvm/NvmContext.tsx
@@ -1,26 +1,25 @@
-import { PropsWithChildren, useEffect, useState } from "react";
+import { PropsWithChildren, useCallback, useEffect, useState } from "react";
 import { createCtx } from "..";
-import { Account, AssetAttributes, AssetPrice, Nevermined, NeverminedOptions } from "@nevermined-io/sdk";
+import { Account, Nevermined, NeverminedOptions } from "@nevermined-io/sdk";
 import { JWTPayload, decodeJwt } from "jose";
 import { useAccount, useWalletClient } from "wagmi";
 import { wagmiConfig } from "~~/services/web3/wagmiConfig";
-import { getMetadata } from "./utils";
 
 // State variables only
 type NvmContextState = {
   provider: any;
   payload: JWTPayload;
   publisher: Account;
+  config: NeverminedOptions;
 };
 
 // This interface differentiates from State
 // because it holds any other option or fx
 // that handle the state in some way
 interface NvmContext extends NvmContextState {
-  publishAsset: () => Promise<void>;
+  login: () => Promise<void>;
 }
 
-// TODO make so that the user can switch chains
 const INITIAL_STATE: NvmContextState = {};
 
 const [useContext, NvmContextProvider] = createCtx<NvmContext>("NvmContext");
@@ -28,65 +27,38 @@ const [useContext, NvmContextProvider] = createCtx<NvmContext>("NvmContext");
 export const NvmProvider = ({ children }: PropsWithChildren) => {
   const [state, setState] = useState<NvmContextState>(INITIAL_STATE);
   const [nevermined, setNvm] = useState<Nevermined>();
-  const [config, setNvmConfig] = useState<NeverminedOptions>();
 
   const { data: signer } = useWalletClient();
   const connector = wagmiConfig.connector;
   const { address } = useAccount();
 
   // at connector update, get the provider, initialize nvm and login
-  useEffect(() => {
-    const startNvm = async () => {
-      try {
-        await getProvider();
-        await initSdk();
-        await login();
-        await publishAsset();
-      } catch (error) {
-        console.error(error);
-      }
-    };
 
-    startNvm();
-    // getProvider()
-    // initSdk();
-    // login();
-    // publishAsset(); //That is just for testing. this function should be used in the Confirm component
-  }, [wagmiConfig.connector]);
-
-  const getProvider = async () => {
+  const getProvider = useCallback(async () => {
     try {
       const provider = await connector?.getProvider();
       setState(prevState => ({ ...prevState, provider }));
     } catch (error) {
-      console.error;
+      console.log(error);
     }
-  };
+  }, [connector]);
 
-  const initSdk = async () => {
-    debugger
+  const initSdk = useCallback(async () => {
     if (!signer || !state.provider) return;
     const config: NeverminedOptions = {
-      // The web3 endpoint of the blockchain network to connect to, could be an Infura endpoint, Quicknode, or any other web3 provider
       web3ProviderUri: "https://goerli-rollup.arbitrum.io/rpc",
       web3Provider: state.provider,
-      // web3Provider: provider,
-      // The url of the marketplace api if you connect to one. It could be your own service if you run a Marketplace API
       marketplaceUri: "https://marketplace-api.goerli.nevermined.app",
-      // The url to a Nevermined node. It could be your own if you run a Nevermined Node
       neverminedNodeUri: "https://node.goerli.nevermined.app",
-      // The public address of the above Node
       neverminedNodeAddress: "0x5838B5512cF9f12FE9f2beccB20eb47211F9B0bc",
-      // The url to access the nevermined subgraphs required to query for on-chain events
       graphHttpUri: "https://api.thegraph.com/subgraphs/name/nevermined-io/public",
-      // Folder where are copied the ABIs of the Nevermined Smart Contracts
       artifactsFolder: "http://localhost:3000/artifacts",
       marketplaceAuthToken:
         localStorage.getItem("marketplaceAuthToken") != null
           ? (localStorage.getItem("marketplaceAuthToken") as string)
           : "",
     };
-    setNvmConfig(config);
+    setState(prevState => ({ ...prevState, config }));
     try {
       const sdk: Nevermined = await Nevermined.getInstance(config);
       setNvm(sdk);
@@ -94,9 +66,9 @@ export const NvmProvider = ({ children }: PropsWithChildren) => {
     } catch (error) {
       console.log(error);
     }
-  };
+  }, [signer, state.provider]);
 
-  const login = async () => {
+  const login = useCallback(async () => {
     try {
       if (!nevermined) return;
 
@@ -106,52 +78,23 @@ export const NvmProvider = ({ children }: PropsWithChildren) => {
       const loginResult = await nevermined.services.marketplace.login(clientAssertion);
       localStorage.setItem("marketplaceAuthToken", loginResult);
       const payload = decodeJwt(loginResult);
-      debugger
       setState(prevState => ({ ...prevState, payload, publisher }));
     } catch (error) {
-      console.error(error);
+      console.log(error);
     }
-  };
+  }, [address, nevermined]);
 
-  /**
-   * 
-   * @info That is currently pushing a mock asset
-   * @todo Publish asset inserted by user (pass in all the props)
-   */
-  const publishAsset = async () => {
-    try {
-      debugger
-      const { publisher, payload } = state;
-      const assetPrice = new AssetPrice(publisher.getId(), 0n);
-
-      const metadata = getMetadata();
-      metadata.main.name = `${metadata.main.name} - ${Math.random()}`;
-      metadata.userId = payload?.sub;
-
-      const assetAttributes = AssetAttributes.getInstance({
-        metadata,
-        services: [
-          {
-            serviceType: "access",
-            price: assetPrice,
-          },
-        ],
-        providers: config?.neverminedNodeAddress ? [config?.neverminedNodeAddress] : undefined,
-      });
-      const steps = [];
-      const ddo = await nevermined?.assets.create(assetAttributes, publisher).next(step => steps.push(step));
-      console.log("ddo", ddo);
-    } catch (error) {
-      console.error(error);
-      return;
-    }
-  };
+  useEffect(() => {
+    if (!signer) return;
+    const initNevermined = async () => await Promise.all([getProvider(), initSdk()]);
+    initNevermined();
+  }, [getProvider, initSdk, signer]);
 
   return (
     <NvmContextProvider
       value={{
         ...state,
-        publishAsset,
+        login,
       }}
     >
       {children}
